# Unmarshal and check printable characters at the same time