--- a/pkg/security/secl/model/utils_test.go
+++ b/pkg/security/secl/model/utils_test.go
@@ -6,6 +6,7 @@
 package model
 
 import (
+	"runtime"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -19,4 +20,13 @@ func TestUnmarshalString(t *testing.T) {
 	}
 
 	assert.Equal(t, "ABC", str)
+}
+
+func BenchmarkUnmarshalPrintableString(b *testing.B) {
+	array := []byte{65, 66, 67, 0, 0, 0, 65, 66}
+	var s string
+	for i := 0; i < b.N; i++ {
+		s, _ = UnmarshalPrintableString(array, len(array))
+	}
+	runtime.KeepAlive(s)
 }