<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Draft PRs should be prefixed with `[WIP]` in their title.

-->
### What does this PR do?
Before
```
BenchmarkUnmarshalPrintableString-4   	 9106581	       119.8 ns/op	      51 B/op	       2 allocs/op
BenchmarkUnmarshalPrintableString-4   	10206292	       121.3 ns/op	      51 B/op	       2 allocs/op
BenchmarkUnmarshalPrintableString-4   	 8898092	       119.5 ns/op	      51 B/op	       2 allocs/op
BenchmarkUnmarshalPrintableString-4   	 9442461	       119.9 ns/op	      51 B/op	       2 allocs/op
BenchmarkUnmarshalPrintableString-4   	10358054	       136.4 ns/op	      51 B/op	       2 allocs/op
```
After
```
BenchmarkUnmarshalPrintableString-4   	30590598	        40.14 ns/op	       3 B/op	       1 allocs/op
BenchmarkUnmarshalPrintableString-4   	30784341	        39.41 ns/op	       3 B/op	       1 allocs/op
BenchmarkUnmarshalPrintableString-4   	30611694	        39.14 ns/op	       3 B/op	       1 allocs/op
BenchmarkUnmarshalPrintableString-4   	30692652	        39.23 ns/op	       3 B/op	       1 allocs/op
BenchmarkUnmarshalPrintableString-4   	30628705	        39.70 ns/op	       3 B/op	       1 allocs/op
```

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [x] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [x] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [x] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [x] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.